move Config to config mod
