+++ frame/abstract/client-s2s/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod client;
+pub mod config;
 pub mod convert;
 pub mod error;
 pub mod types {
