+++ frame/assistants/relay-s2s/src/message/receiving_relay.rs
@@ -1,4 +1,5 @@
-use abstract_client_s2s::client::{Config, S2SClientRelay};
+use abstract_client_s2s::client::S2SClientRelay;
+use abstract_client_s2s::config::Config;
 use abstract_client_s2s::convert::SmartCodecMapper;
 use abstract_client_s2s::types::bp_messages::{OutboundLaneData, UnrewardedRelayersState};
 use abstract_client_s2s::types::bridge_runtime_common::messages::source::FromBridgedChainMessagesDeliveryProof;
